+++ src/commands/cmd_key.cc
@@ -19,8 +19,8 @@
  */
 
 #include "commands/ttl_util.h"
-#include "redis_cmd.h"
-#include "redis_cmd_error.h"
+#include "commander.h"
+#include "error_constants.h"
 #include "server/server.h"
 #include "storage/redis_db.h"
 #include "time_util.h"
