+++ src/commands/cmd_pubsub.cc
@@ -18,8 +18,8 @@
  *
  */
 
-#include "redis_cmd.h"
-#include "redis_cmd_error.h"
+#include "commander.h"
+#include "error_constants.h"
 #include "server/server.h"
 #include "storage/redis_pubsub.h"
