+++ src/commands/cmd_server.cc
@@ -18,10 +18,10 @@
  *
  */
 
-#include "commands/scan_util.h"
+#include "commands/scan_base.h"
 #include "config/config.h"
-#include "redis_cmd.h"
-#include "redis_cmd_error.h"
+#include "commander.h"
+#include "error_constants.h"
 #include "server/redis_connection.h"
 #include "server/server.h"
 #include "stats/disk_stats.h"
