+++ src/commands/cmd_set.cc
@@ -18,9 +18,9 @@
  *
  */
 
-#include "commands/scan_util.h"
-#include "redis_cmd.h"
-#include "redis_cmd_error.h"
+#include "commands/scan_base.h"
+#include "commander.h"
+#include "error_constants.h"
 #include "server/server.h"
 #include "types/redis_set.h"
