+++ src/commands/cmd_sortedint.cc
@@ -18,8 +18,8 @@
  *
  */
 
-#include "redis_cmd.h"
-#include "redis_cmd_error.h"
+#include "commander.h"
+#include "error_constants.h"
 #include "server/server.h"
 #include "types/redis_sortedint.h"
