+++ src/commands/cmd_txn.cc
@@ -18,8 +18,8 @@
  *
  */
 
-#include "redis_cmd.h"
-#include "redis_cmd_error.h"
+#include "commander.h"
+#include "error_constants.h"
 #include "server/redis_connection.h"
 
 namespace Redis {
